+++ packages/firebase_auth/firebase_auth/test/firebase_auth_test.dart
@@ -10,6 +10,7 @@ import 'package:firebase_auth_platform_interface/firebase_auth_platform_interfac
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
+import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 
 const String kMockProviderId = 'firebase';
 const String kMockUid = '12345';
@@ -28,6 +29,7 @@ const String kMockVerificationId = '12345';
 const String kMockSmsCode = '123456';
 const String kMockLanguage = 'en';
 const String kMockIdTokenResultSignInProvider = 'password';
+const String kMockOobCode = 'oobcode';
 const Map<dynamic, dynamic> kMockIdTokenResultClaims = <dynamic, dynamic>{
   'claim1': 'value1',
 };
@@ -87,7 +89,6 @@ void main() {
 
     setUp(() {
       mock = MockFirebaseAuth();
-      when(mock.isMock).thenReturn(true);
       when(mock.getIdToken(any, any)).thenAnswer(
           (_) => Future<PlatformIdTokenResult>.value(kMockIdTokenResult));
       when(mock.isSignInWithEmailLink(any, any))
@@ -876,7 +877,15 @@ void main() {
       await auth.setLanguageCode(kMockLanguage);
       verify(mock.setLanguageCode(auth.app.name, kMockLanguage));
     });
+
+    test('confirmPasswordReset', () async {
+      await auth.confirmPasswordReset(kMockOobCode, kMockPassword);
+      verify(mock.confirmPasswordReset(
+          auth.app.name, kMockOobCode, kMockPassword));
+    });
   });
 }
 
-class MockFirebaseAuth extends Mock implements FirebaseAuthPlatform {}
+class MockFirebaseAuth extends Mock
+    with MockPlatformInterfaceMixin
+    implements FirebaseAuthPlatform {}