+++ packages/cloud_firestore/cloud_firestore/CHANGELOG.md
@@ -1,3 +1,46 @@
+## 0.13.4
+
+* Support equality comparison on `FieldValue` instances.
+* Updated version of endorsed web implementation.
+
+## 0.13.3+1
+
+* Make the pedantic dev_dependency explicit.
+
+## 0.13.3
+
+* Add macOS support
+
+## 0.13.2+3
+
+* Fixed decoding & encoding platform interface instances in nested maps
+
+## 0.13.2+2
+
+* Fixed crashes when using `FieldValue.arrayUnion` & `FieldValue.arrayRemove` with `DocumentReference` objects
+
+## 0.13.2+1
+
+* Add Web integration documentation to README.
+
+## 0.13.2
+
+* Add web support by default.
+* Require Flutter SDK 1.12.13+hotfix.4 or later
+* Add web support to the example app.
+
+## 0.13.1+1
+
+* Fixed crashes when using `Query#where` with `DocumentReference` objects
+
+## 0.13.1
+
+* Migrate to `cloud_firestore_platform_interface`.
+
+## 0.13.0+2
+
+* Fixed `persistenceEnabled`, `sslEnabled`, and `timestampsInSnapshotsEnabled` on iOS.
+
 ## 0.13.0+1
 
 * Remove the deprecated `author:` field from pubspec.yaml