+++ packages/firebase_admob/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.9.1+3
+
+* Replace deprecated `getFlutterEngine` call on Android.
+
+## 0.9.1+2
+
+* Make the pedantic dev_dependency explicit.
+
+## 0.9.1+1
+
+* Enable custom parameters for rewarded video server-side verification callbacks.
+
 ## 0.9.1
 
 * Support v2 embedding. This will remain compatible with the original embedding and won't require