+++ packages/firebase_auth/firebase_auth_web/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.1.2+2
+
+* Make the pedantic dev_dependency explicit.
+
+## 0.1.2+1
+
+* Require `firebase_core_web` from hosted
+
+## 0.1.2
+
+* Implement `fetchSignInMethodsForEmail`, `isSignInWithEmailLink`, `signInWithEmailAndLink`, and `sendLinkToEmail`. 
+
 ## 0.1.1+4
 
 * Prevent `null` users (unauthenticated) from breaking the `onAuthStateChanged` Stream.