+++ packages/firebase_crashlytics/CHANGELOG.md
@@ -1,3 +1,16 @@
+## 0.1.3+1
+
+* Make the pedantic dev_dependency explicit.
+
+## 0.1.3
+
+* Add macOS support
+
+## 0.1.2+5
+
+* Fix overrides a deprecated API.
+* Raise minimum required Flutter SDK version to 1.12.13+hotfix.4
+
 ## 0.1.2+4
 
 * Updated the example with the missing `recordError()` method.