+++ packages/firebase_dynamic_links/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.5.0+11
+
+* Replace deprecated `getFlutterEngine` call on Android.
+
+## 0.5.0+10
+
+* Make the pedantic dev_dependency explicit.
+
 ## 0.5.0+9
 
 * Remove the deprecated `author:` field from pubspec.yaml