+++ packages/firebase_messaging/CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 6.1.0
+
 * Implement `UNUserNotificationCenterDelegate` methods to allow plugin to work when method swizzling is disabled
 * **Breaking change** Applications now only need to their iOS project's `AppDelegate` when method swizzling is disabled
 * **Breaking change** Applications that need to use `firebase_messaging` with other notification plugins will
@@ -7,6 +8,14 @@
   <key>FirebaseAppDelegateProxyEnabled</key>
   <false/>
   ```
+
+## 6.0.12
+
+* Replace deprecated `getFlutterEngine` call on Android.
+
+## 6.0.11
+
+* Make the pedantic dev_dependency explicit.
 
 ## 6.0.10