+++ packages/firebase_performance/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.3.1+7
+
+* Replace deprecated `getFlutterEngine` call on Android.
+
+## 0.3.1+6
+
+* Make the pedantic dev_dependency explicit.
+
 ## 0.3.1+5
 
 * Remove the deprecated `author:` field from pubspec.yaml