+++ packages/firebase_storage/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 3.1.3
+
+* Replace deprecated `getFlutterEngine` call on Android.
+
+## 3.1.2
+
+* Make the pedantic dev_dependency explicit.
+
 ## 3.1.1
 
 * Removed unnecessary debug print statements ("i am working").